# website/openapi/engine.openapi.json
{
  "components": {
    "schemas": {
      "DefinitionActiveUpdate": {
        "description": "PATCH body to soft-enable or soft-disable a catalog definition.",
        "properties": {
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          }
        },
        "required": [
          "is_active"
        ],
        "title": "DefinitionActiveUpdate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SagaDefinitionItem": {
        "description": "One row in GET /v1/definitions/sagas.",
        "properties": {
          "body": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Full manifest blueprint when include_body=true on by-id GET.",
            "title": "Body"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "description": "Saga definition UUID",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "namespace",
          "name",
          "version",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "SagaDefinitionItem",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Control-plane HTTP API for deploying manifests, starting sagas, HITL review, and operator recovery. OSS surface under /v1.",
    "title": "API Reference",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/definitions/sagas": {
      "get": {
        "description": "List saga definitions, or fetch one by ``id`` XOR ``namespace+name+version``.",
        "operationId": "get_definitions_sagas_v1_definitions_sagas_get",
        "parameters": [
          {
            "description": "Definition UUID for single-item fetch (XOR with triple).",
            "in": "query",
            "name": "id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Definition UUID for single-item fetch (XOR with triple).",
              "title": "Id"
            }
          },
          {
            "description": "Filter or triple identity.",
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter or triple identity.",
              "title": "Namespace"
            }
          },
          {
            "description": "Exact name filter or triple identity.",
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact name filter or triple identity.",
              "title": "Name"
            }
          },
          {
            "description": "With namespace+name: single-item fetch by triple.",
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "With namespace+name: single-item fetch by triple.",
              "title": "Version"
            }
          },
          {
            "description": "Filter by active flag (list mode).",
            "in": "query",
            "name": "is_active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by active flag (list mode).",
              "title": "Is Active"
            }
          },
          {
            "description": "When fetching one definition, include full manifest body.",
            "in": "query",
            "name": "include_body",
            "required": false,
            "schema": {
              "default": false,
              "description": "When fetching one definition, include full manifest body.",
              "title": "Include Body",
              "type": "boolean"
            }
          },
          {
            "description": "Include total matching row count.",
            "in": "query",
            "name": "include_total",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include total matching row count.",
              "title": "Include Total",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SagaDefinitionItem"
                    },
                    {
                      "$ref": "#/components/schemas/SagaDefinitionListResponse"
                    }
                  ],
                  "title": "Response Get Definitions Sagas V1 Definitions Sagas Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Definitions Sagas",
        "tags": [
          "definitions"
        ]
      },
      "patch": {
        "description": "Soft-enable or soft-disable a saga definition (`is_active`) by id or triple.",
        "operationId": "patch_definitions_saga_active_v1_definitions_sagas_patch",
        "parameters": [
          {
            "description": "Definition UUID (mutually exclusive with triple).",
            "in": "query",
            "name": "id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Definition UUID (mutually exclusive with triple).",
              "title": "Id"
            }
          },
          {
            "description": "With name+version: identity triple.",
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "With name+version: identity triple.",
              "title": "Namespace"
            }
          },
          {
            "description": "With namespace+version: identity triple.",
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "With namespace+version: identity triple.",
              "title": "Name"
            }
          },
          {
            "description": "With namespace+name: identity triple.",
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "With namespace+name: identity triple.",
              "title": "Version"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefinitionActiveUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SagaDefinitionItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Definitions Saga Active",
        "tags": [
          "definitions"
        ]
      }
    },
    "/v1/definitions/sagas/{definition_id}": {
      "get": {
        "description": "Return one saga definition by primary key UUID (for start-saga resolution).",
        "operationId": "get_definitions_saga_by_id_v1_definitions_sagas__definition_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          },
          {
            "description": "When true, include the full manifest blueprint in body.",
            "in": "query",
            "name": "include_body",
            "required": false,
            "schema": {
              "default": false,
              "description": "When true, include the full manifest blueprint in body.",
              "title": "Include Body",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SagaDefinitionItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Definitions Saga By Id",
        "tags": [
          "definitions"
        ]
      }
    },
...
    "/v1/definitions/steps/{definition_id}": {
      "get": {
        "description": "Return one step definition by primary key UUID.",
        "operationId": "get_definitions_step_by_id_v1_definitions_steps__definition_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          },
          {
            "description": "When true, include the full step manifest blueprint in body.",
            "in": "query",
            "name": "include_body",
            "required": false,
            "schema": {
              "default": false,
              "description": "When true, include the full step manifest blueprint in body.",
              "title": "Include Body",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepDefinitionItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Definitions Step By Id",
        "tags": [
          "definitions"
        ]
      }
    },
...
    "/v1/definitions/workers/{definition_id}": {
      "get": {
        "description": "Return one worker definition by primary key UUID.",
        "operationId": "get_definitions_worker_by_id_v1_definitions_workers__definition_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          },
          {
            "description": "When true, include the full manifest body.",
            "in": "query",
            "name": "include_body",
            "required": false,
            "schema": {
              "default": false,
              "description": "When true, include the full manifest body.",
              "title": "Include Body",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerDefinitionItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Definitions Worker By Id",
        "tags": [
          "definitions"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local engine",
      "url": "http://127.0.0.1:8000"
    }
  ]
}